Rename utils module and update embeds
